Refactor getPartitionPath to MetadataPartitionType
